add admin in database with bcrypt
