+++ package.json
@@ -9,7 +9,9 @@
   "author": "JOHAN",
   "license": "ISC",
   "dependencies": {
+    "bcrypt": "^5.1.1",
     "body-parser": "^1.20.2",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "mysql2": "^3.6.0",
     "serve-favicon": "^2.5.0"
